+++ server/src/formatter/formatterDetail.ts
@@ -1,6 +1,6 @@
 import {Position} from "vscode-languageserver";
 import {FormatterState, stepCursorAlongLines} from "./formatterState";
-import {TokenObject, TokenKind} from "../compiler_tokenizer/tokens";
+import {TokenObject, TokenKind} from "../compiler_tokenizer/tokenObject";
 import {NodesBase} from "../compiler_parser/nodes";
 import {tracer} from "../code/tracer";
 import {getGlobalSettings} from "../code/settings";
